Separate workflow command parsing from runtime execution
Amp-Thread-ID: https://ampcode.com/threads/T-01a0606f-6a04-757c-84b0-caf85c154ab0
